Refactor: useChat chatting custom hook (#62)
